requirements-test.txt
# Testing Framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0
pytest-timeout>=2.1.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0
pytest-randomly>=3.14.0

# Coverage Analysis
coverage>=7.3.0
coverage-badge>=1.1.0

# Test Utilities
mock>=5.1.0
faker>=19.3.0

# Code Quality & Linting
flake8>=6.0.0
pylint>=2.17.0
black>=23.7.0
isort>=5.12.0
autopep8>=2.0.0

# Security Scanning
bandit>=1.7.5

# Performance & Benchmarking
memory-profiler>=0.61.0
pytest-benchmark>=4.0.0
pytest-profiling>=1.7.0

# Type Checking
mypy>=1.4.0

# Test Reporting
pytest-html>=3.2.0
pytest-json-report>=1.5.0

# Utilities
pyyaml>=6.0
